Update marcap_utils.py
윈도우(python: 3.7.6, pandas: 1.2.4, 1.2.5 둘 모두)에서 pd.concat()에서 에러 발생  #6 
Amount, Marcap 타입을 int에서 float로 수정

+++ marcap_utils.py
@@ -18,11 +18,11 @@ def marcap_data(start, end=None, code=None):
   df_list = []
 
   dtypes={'Code':str, 'Name':str, 
-          'Open':int, 'High':int, 'Low':int, 'Close':int, 'Volume':int, 'Amount':int,
-          'Changes':int, 'ChangeCode':str, 'ChagesRatio':float, 'Marcap':int, 'Stocks':int,
+          'Open':int, 'High':int, 'Low':int, 'Close':int, 'Volume':int, 'Amount':float,
+          'Changes':int, 'ChangeCode':str, 'ChagesRatio':float, 'Marcap':float, 'Stocks':int,
           'MarketId':str, 'Market':str, 'Dept':str,
           'Rank':int}
-
+    
   for year in range(start.year, end.year + 1):
     try:
       csv_file = 'marcap/data/marcap-%s.csv.gz' % (year)
